pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "casablanca-builtup-pipeline"
version = "0.1.0"
description = "Local production-grade Sentinel-1/Sentinel-2 built-up change detection pipeline for Casablanca."
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
  "boto3>=1.35.0",
  "click>=8.1.7,<8.2",
  "dask[array]>=2024.8.0",
  "geopandas>=0.14.4",
  "numpy>=1.26.4",
  "odc-geo>=0.4.8",
  "odc-stac>=0.3.11",
  "pandas>=2.2.2",
  "planetary-computer>=1.0.0",
  "pydantic>=2.8.2",
  "pyarrow>=17.0.0",
  "pyogrio>=0.9.0",
  "pystac>=1.10.1",
  "pystac-client>=0.8.3",
  "PyYAML>=6.0.2",
  "rasterio>=1.3.10",
  "rioxarray>=0.17.0",
  "scikit-image>=0.24.0",
  "scipy>=1.13.1",
  "shapely>=2.0.5",
  "typer>=0.12.3",
  "xarray>=2024.7.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "pytest-cov>=5.0.0",
]

[project.scripts]
casablanca-builtup = "casablanca_builtup.cli:app"
casablanca-psi = "casablanca_psi.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
